+++ WebContent/template/top.jsp
@@ -1,6 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=EUC-KR"
 	pageEncoding="EUC-KR"%>
-<%@page import="emp.dto.LoginDTO"%>
+<%@page import="erp.dto.LoginDTO"%>
 	
 <!DOCTYPE html>
 <html>
