+++ src/erp/controller/EmpUpdateServlet.java
@@ -1,4 +1,4 @@
-package emp.controller;
+package erp.controller;
 
 import java.io.IOException;
 
@@ -9,9 +9,9 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import emp.dto.EmpDTO;
-import emp.service.EmpService;
-import emp.service.EmpServiceImpl;
+import erp.dto.EmpDTO;
+import erp.service.EmpService;
+import erp.service.EmpServiceImpl;
 @WebServlet(name = "emp/update", urlPatterns = { "/emp/update.do" })
 public class EmpUpdateServlet extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
