+++ src/erp/dto/EmpDTO.java
@@ -1,4 +1,4 @@
-package emp.dto;
+package erp.dto;
 
 import java.sql.Date;
