+++ src/erp/dto/MemberDTO.java
@@ -1,9 +1,8 @@
-package emp.dto;
+package erp.dto;
 
 import java.sql.Date;
 
-public class LoginDTO {
-
+public class MemberDTO {
 	private String id;
 	private String pass;
 	private String name;
@@ -26,18 +25,13 @@ public class LoginDTO {
 	private String phonecell;
 	private String email;
 	private String profile_photo;
-	private String deptname;
-	private String job_category;
-	private String menupath;
-
-	LoginDTO() {
-
+	public MemberDTO(){
+		
 	}
-
-	public LoginDTO(String id, String pass, String name, String ssn, Date birthday, String marry, String gender,
+	public MemberDTO(String id, String pass, String name, String ssn, Date birthday, String marry, String gender,
 			String position, String duty, String classes, Date startday, Date endday, String deptno, String curstate,
 			String zipcode, String addr, String detailaddr, String phonehome, String phoneco, String phonecell,
-			String email, String profile_photo, String deptname, String job_category, String menupath) {
+			String email, String profile_photo) {
 		super();
 		this.id = id;
 		this.pass = pass;
@@ -61,222 +55,149 @@ public class LoginDTO {
 		this.phonecell = phonecell;
 		this.email = email;
 		this.profile_photo = profile_photo;
-		this.deptname = deptname;
-		this.job_category = job_category;
-		this.menupath = menupath;
 	}
-
+	public String getId() {
+		return id;
+	}
+	public void setId(String id) {
+		this.id = id;
+	}
+	public String getPass() {
+		return pass;
+	}
+	public void setPass(String pass) {
+		this.pass = pass;
+	}
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	public String getSsn() {
+		return ssn;
+	}
+	public void setSsn(String ssn) {
+		this.ssn = ssn;
+	}
+	public Date getBirthday() {
+		return birthday;
+	}
+	public void setBirthday(Date birthday) {
+		this.birthday = birthday;
+	}
+	public String getMarry() {
+		return marry;
+	}
+	public void setMarry(String marry) {
+		this.marry = marry;
+	}
+	public String getGender() {
+		return gender;
+	}
+	public void setGender(String gender) {
+		this.gender = gender;
+	}
+	public String getPosition() {
+		return position;
+	}
+	public void setPosition(String position) {
+		this.position = position;
+	}
+	public String getDuty() {
+		return duty;
+	}
+	public void setDuty(String duty) {
+		this.duty = duty;
+	}
+	public String getClasses() {
+		return classes;
+	}
+	public void setClasses(String classes) {
+		this.classes = classes;
+	}
+	public Date getStartday() {
+		return startday;
+	}
+	public void setStartday(Date startday) {
+		this.startday = startday;
+	}
+	public Date getEndday() {
+		return endday;
+	}
+	public void setEndday(Date endday) {
+		this.endday = endday;
+	}
+	public String getDeptno() {
+		return deptno;
+	}
+	public void setDeptno(String deptno) {
+		this.deptno = deptno;
+	}
+	public String getCurstate() {
+		return curstate;
+	}
+	public void setCurstate(String curstate) {
+		this.curstate = curstate;
+	}
+	public String getZipcode() {
+		return zipcode;
+	}
+	public void setZipcode(String zipcode) {
+		this.zipcode = zipcode;
+	}
+	public String getAddr() {
+		return addr;
+	}
+	public void setAddr(String addr) {
+		this.addr = addr;
+	}
+	public String getDetailaddr() {
+		return detailaddr;
+	}
+	public void setDetailaddr(String detailaddr) {
+		this.detailaddr = detailaddr;
+	}
+	public String getPhonehome() {
+		return phonehome;
+	}
+	public void setPhonehome(String phonehome) {
+		this.phonehome = phonehome;
+	}
+	public String getPhoneco() {
+		return phoneco;
+	}
+	public void setPhoneco(String phoneco) {
+		this.phoneco = phoneco;
+	}
+	public String getPhonecell() {
+		return phonecell;
+	}
+	public void setPhonecell(String phonecell) {
+		this.phonecell = phonecell;
+	}
+	public String getEmail() {
+		return email;
+	}
+	public void setEmail(String email) {
+		this.email = email;
+	}
+	public String getProfile_photo() {
+		return profile_photo;
+	}
+	public void setProfile_photo(String profile_photo) {
+		this.profile_photo = profile_photo;
+	}
 	@Override
 	public String toString() {
-		return "LoginDTO [id=" + id + ", pass=" + pass + ", name=" + name + ", ssn=" + ssn + ", birthday=" + birthday
+		return "MemberDTO [id=" + id + ", pass=" + pass + ", name=" + name + ", ssn=" + ssn + ", birthday=" + birthday
 				+ ", marry=" + marry + ", gender=" + gender + ", position=" + position + ", duty=" + duty + ", classes="
 				+ classes + ", startday=" + startday + ", endday=" + endday + ", deptno=" + deptno + ", curstate="
 				+ curstate + ", zipcode=" + zipcode + ", addr=" + addr + ", detailaddr=" + detailaddr + ", phonehome="
 				+ phonehome + ", phoneco=" + phoneco + ", phonecell=" + phonecell + ", email=" + email
-				+ ", profile_photo=" + profile_photo + ", deptname=" + deptname + ", job_category=" + job_category
-				+ ", menupath=" + menupath + "]";
+				+ ", profile_photo=" + profile_photo + "]";
 	}
-
-	public String getId() {
-		return id;
-	}
-
-	public void setId(String id) {
-		this.id = id;
-	}
-
-	public String getPass() {
-		return pass;
-	}
-
-	public void setPass(String pass) {
-		this.pass = pass;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getSsn() {
-		return ssn;
-	}
-
-	public void setSsn(String ssn) {
-		this.ssn = ssn;
-	}
-
-	public Date getBirthday() {
-		return birthday;
-	}
-
-	public void setBirthday(Date birthday) {
-		this.birthday = birthday;
-	}
-
-	public String getMarry() {
-		return marry;
-	}
-
-	public void setMarry(String marry) {
-		this.marry = marry;
-	}
-
-	public String getGender() {
-		return gender;
-	}
-
-	public void setGender(String gender) {
-		this.gender = gender;
-	}
-
-	public String getPosition() {
-		return position;
-	}
-
-	public void setPosition(String position) {
-		this.position = position;
-	}
-
-	public String getDuty() {
-		return duty;
-	}
-
-	public void setDuty(String duty) {
-		this.duty = duty;
-	}
-
-	public String getClasses() {
-		return classes;
-	}
-
-	public void setClasses(String classes) {
-		this.classes = classes;
-	}
-
-	public Date getStartday() {
-		return startday;
-	}
-
-	public void setStartday(Date startday) {
-		this.startday = startday;
-	}
-
-	public Date getEndday() {
-		return endday;
-	}
-
-	public void setEndday(Date endday) {
-		this.endday = endday;
-	}
-
-	public String getDeptno() {
-		return deptno;
-	}
-
-	public void setDeptno(String deptno) {
-		this.deptno = deptno;
-	}
-
-	public String getCurstate() {
-		return curstate;
-	}
-
-	public void setCurstate(String curstate) {
-		this.curstate = curstate;
-	}
-
-	public String getZipcode() {
-		return zipcode;
-	}
-
-	public void setZipcode(String zipcode) {
-		this.zipcode = zipcode;
-	}
-
-	public String getAddr() {
-		return addr;
-	}
-
-	public void setAddr(String addr) {
-		this.addr = addr;
-	}
-
-	public String getDetailaddr() {
-		return detailaddr;
-	}
-
-	public void setDetailaddr(String detailaddr) {
-		this.detailaddr = detailaddr;
-	}
-
-	public String getPhonehome() {
-		return phonehome;
-	}
-
-	public void setPhonehome(String phonehome) {
-		this.phonehome = phonehome;
-	}
-
-	public String getPhoneco() {
-		return phoneco;
-	}
-
-	public void setPhoneco(String phoneco) {
-		this.phoneco = phoneco;
-	}
-
-	public String getPhonecell() {
-		return phonecell;
-	}
-
-	public void setPhonecell(String phonecell) {
-		this.phonecell = phonecell;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getProfile_photo() {
-		return profile_photo;
-	}
-
-	public void setProfile_photo(String profile_photo) {
-		this.profile_photo = profile_photo;
-	}
-
-	public String getDeptname() {
-		return deptname;
-	}
-
-	public void setDeptname(String deptname) {
-		this.deptname = deptname;
-	}
-
-	public String getJob_category() {
-		return job_category;
-	}
-
-	public void setJob_category(String job_category) {
-		this.job_category = job_category;
-	}
-
-	public String getMenupath() {
-		return menupath;
-	}
-
-	public void setMenupath(String menupath) {
-		this.menupath = menupath;
-	}
-
-	
 	
 }
+
+
