+++ src/query/EmpQuery.java
@@ -26,9 +26,6 @@ public class EmpQuery {
 	public static final String UPDATE_EMP
 	="update kitriemp set addr=?, point=?, grade=? where id=?";
 	
-/*	public static final String LOGIN
-	="select * from kitriemp where id=? and pass=?";*/
-	
 	public static final String LOGIN
 	="select m.*, d.deptname, j.job_category, j.menupath from member m, dept d, "
 			+ "job j where m.deptno = d.deptno and d.job_category = j.job_id and id = ? and pass = ?";
